separate high and low level interfaces, including making low level inteface 100% deterministic, passing in random seeds when needed from high level interface
